requirements.txt
openai==1.54.0
pydantic==2.9.0
python-dotenv==1.0.0
pyyaml==6.0.2
dateparser>=1.2.0
python-dateutil>=2.8.2
anthropic>=0.40.0

# Web Server & API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.28.0
